requirements.txt
absl-py==1.1.0
albumentations==1.1.0
attrs==21.4.0
cachetools==5.2.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
cycler==0.11.0
fonttools==4.33.3
google-auth==2.6.6
google-auth-oauthlib==0.4.6
grpcio==1.46.3
idna==3.3
imageio==2.19.3
importlib-metadata==4.11.4
iniconfig==1.1.1
joblib==1.1.0
kiwisolver==1.4.2
Markdown==3.3.7
matplotlib==3.5.2
networkx==2.8.3
numpy==1.22.4
oauthlib==3.2.0
opencv-python-headless==4.5.5.64
packaging==21.3
Pillow==9.1.1
pluggy==1.0.0
protobuf==3.20.1
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
pytest==7.1.2
python-dateutil==2.8.2
python-dotenv==0.20.0
PyWavelets==1.3.0
PyYAML==6.0
qudida==0.0.4
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-image==0.19.2
scikit-learn==1.1.1
scipy==1.8.1
six==1.16.0
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
tifffile==2022.5.4
tomli==2.0.1
torch==1.11.0+cu113
torchaudio
torchinfo==1.7.0
torchvision
tqdm==4.64.0
typing_extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0
